+++ modules/libsbc/CMakeLists.txt
@@ -1,11 +1,10 @@
-if(CONFIG_LIBSBC_ENCODER OR CONFIG_LIBSBC_DECODER)
+if(CONFIG_LIBSBC)
 
 zephyr_library_named(libsbc)
 zephyr_library_compile_options(-O3 -std=c11 -ffast-math -Wno-stringop-overflow)
 
 zephyr_compile_definitions(SBC_FOR_EMBEDDED_LINUX)
 zephyr_compile_definitions(SBC_NO_PCM_CPY_OPTION)
-zephyr_library_sources(sbc.c)
 
 zephyr_include_directories(
     ${ZEPHYR_LIBSBC_MODULE_DIR}/encoder/include
@@ -16,7 +15,6 @@ zephyr_include_directories(
   ${ZEPHYR_LIBSBC_MODULE_DIR}/decoder/srce
   )
 
-if(CONFIG_LIBSBC_ENCODER)
 zephyr_library_sources(
   ${ZEPHYR_LIBSBC_MODULE_DIR}/encoder/srce/sbc_analysis.c
   ${ZEPHYR_LIBSBC_MODULE_DIR}/encoder/srce/sbc_dct.c
@@ -27,9 +25,7 @@ zephyr_library_sources(
   ${ZEPHYR_LIBSBC_MODULE_DIR}/encoder/srce/sbc_encoder.c
   ${ZEPHYR_LIBSBC_MODULE_DIR}/encoder/srce/sbc_packing.c
   )
-endif()
 
-if(CONFIG_LIBSBC_DECODER)
 zephyr_library_sources(
   ${ZEPHYR_LIBSBC_MODULE_DIR}/decoder/srce/alloc.c
   ${ZEPHYR_LIBSBC_MODULE_DIR}/decoder/srce/bitalloc.c
@@ -47,5 +43,4 @@ zephyr_library_sources(
   ${ZEPHYR_LIBSBC_MODULE_DIR}/decoder/srce/synthesis-dct8.c
   ${ZEPHYR_LIBSBC_MODULE_DIR}/decoder/srce/synthesis-sbc.c
   )
-endif()
 endif()
